group durableClient files together
