+++ src/durableClient/PurgeHistoryResult.ts
@@ -1,0 +1,5 @@
+import * as types from "durable-functions";
+
+export class PurgeHistoryResult implements types.PurgeHistoryResult {
+    constructor(public readonly instancesDeleted: number) {}
+}
