+++ src/index.ts
@@ -1,10 +1,10 @@
-import { getClient } from "./durableorchestrationclient";
 import { DummyEntityContext, DummyOrchestrationContext } from "./util/testingUtils";
 import { ManagedIdentityTokenSource } from "./tokensource";
 import { EntityId } from "./entities/entityid";
 import { EntityStateResponse } from "./entities/entitystateresponse";
 import { OrchestrationRuntimeStatus } from "./orchestrations/OrchestrationRuntimeStatus";
 import { RetryOptions } from "./retryoptions";
+import { getClient } from "./durableClient/getClient";
 
 export * as app from "./app";
 export * as trigger from "./trigger";
